refactor: 인터페이스 파일명을 type.ts로 정리
