+++ src/widgets/post/ui/EditPostDialog.tsx
@@ -1,4 +1,4 @@
-import { Post } from "../../../entities/post/model/post"
+import { Post } from "../../../entities/post/model/type"
 import { Button } from "../../../shared/ui/Button"
 import {
   Dialog,
